subo corregido

+++ ft_strdup.c
@@ -3,16 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strdup.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alex <alex@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: alcarril <alcarril@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/28 16:26:30 by alcarril          #+#    #+#             */
-/*   Updated: 2024/10/03 23:29:24 by alex             ###   ########.fr       */
+/*   Updated: 2024/10/04 13:41:55 by alcarril         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 char	*ft_strdup(const char *s)
 {
-	int		i;
+	size_t		i;
 	char	*ptr;
 	size_t	num_values;
